All float types have been switcehd to double

+++ HELLION.DataStructures/HECelestialBody.cs
@@ -8,19 +8,19 @@ namespace HELLION.DataStructures
         public string Name { get; set; }
         public long ParentGUID { get; set; }
         public double Mass { get; set; }
-        public float Radius { get; set; }
-        public float RotationPeriod { get; set; }
-        public float Eccentricity { get; set; }
-        public float SemiMajorAxis { get; set; }
-        public float Inclination { get; set; }
-        public float ArgumentOfPeriapsis { get; set; }
-        public float LongitudeOfAscendingNode { get; set; }
+        public double Radius { get; set; }
+        public double RotationPeriod { get; set; }
+        public double Eccentricity { get; set; }
+        public double SemiMajorAxis { get; set; }
+        public double Inclination { get; set; }
+        public double ArgumentOfPeriapsis { get; set; }
+        public double LongitudeOfAscendingNode { get; set; }
         public string PlanetsPrefabPath { get; set; }
         public string MainCameraPrefabPath { get; set; }
         public string NavigationPrefabPath { get; set; }
-        public float AtmosphereLevel1 { get; set; }
-        public float AtmosphereLevel2 { get; set; }
-        public float AtmosphereLevel3 { get; set; }
-        public float ResourceResolution { get; set; }
+        public double AtmosphereLevel1 { get; set; }
+        public double AtmosphereLevel2 { get; set; }
+        public double AtmosphereLevel3 { get; set; }
+        public double ResourceResolution { get; set; }
     }
 } // End of namespace HELLION.DataStructures
